comments to what is ethereum chapter
Some suggestions on how to make this intro more complete.
As a side comment for the rest of the book, I would not be shy talking about bitcoin directly (instead of referring to hypothetical blockchains) since everyone knows Andreas’ expertise and passion about bitcoin and nobody would think that it’s not politically correct to use the world bitcoin. It would actually perfectly fit with Andreas’ intellectual honesty and possibly help the community to mature.